fix the bug for intersect of node and obstacles by modifying the type of position of node from int to float in constructor and float to int in __str__ method

+++ python/pyhpp/node.py
@@ -3,9 +3,9 @@ from math import inf, sqrt
 
 class Node:
     def __init__(self, x, y, z = None):
-        self.x = int(x)
-        self.y = int(y)
-        self.z = None if z is None else int(z)
+        self.x = float(x)
+        self.y = float(y)
+        self.z = None if z is None else float(z)
         self.is_2d = True if z is None else False
 
         self.prev = None
@@ -13,8 +13,8 @@ class Node:
         self.f = inf
 
     def __str__(self):
-        return str(self.x) + ',' + str(self.y) if self.is_2d \
-            else str(self.x) + ',' + str(self.y) + ',' + str(self.z)
+        return str(int(self.x)) + ',' + str(int(self.y)) if self.is_2d \
+            else str(int(self.x)) + ',' + str(int(self.y)) + ',' + str(int(self.z))
 
     def __eq__(self, other):
         return str(self) == str(other)
@@ -31,8 +31,8 @@ class Node:
         self.dist = 0
     
     def shift(self, x, y, z = 0):
-        return Node(self.x + int(x), self.y + int(y)) if self.is_2d \
-            else Node(self.x + int(x), self.y + int(y), self.z + int(z))
+        return Node(self.x + x, self.y + y) if self.is_2d \
+            else Node(self.x + x, self.y + y, self.z + z)
 
     def manhattan_distance_to(self, destNode) -> float:
         distance = abs(destNode.x - self.x) + abs(destNode.y - self.y)
